refactor: 💡Move tests for LocaleService
