+++ apps/api/src/test/java/dk/treecreate/api/mail/MailServiceTest.java
@@ -1,15 +1,12 @@
 package dk.treecreate.api.mail;
 
-import dk.treecreate.api.utils.LocaleService;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.mock.mockito.MockBean;
 
-import java.util.Locale;
 import java.util.stream.Stream;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
@@ -19,8 +16,6 @@ class MailServiceTest
 {
     @Autowired
     MailService mailService;
-    @Autowired
-    private LocaleService localeService;
 
     // sendSignupEmail and sendResetPasswordEmail are not tested due to their simplicity
     // Potential test would be validating that sendMail() method gets called correctly but it is private
@@ -74,27 +69,11 @@ class MailServiceTest
                 "Åæøæøæøæøæøæøæøæøæøæøæøæøæøæ" + "@domain.yahoo", false));
     }
 
-    private static Stream<Arguments> getLocaleArguments()
-    {
-        return Stream.of(
-            Arguments.of("dk", new Locale("dk")),
-            Arguments.of("en", Locale.ENGLISH),
-            Arguments.of(null, new Locale("dk")));
-    }
-
     @ParameterizedTest
     @MethodSource("isValidEmailArguments")
     @DisplayName("isValidEmail() returns correct value")
     void isValidEmail(String email, boolean expected)
     {
         assertEquals(expected, mailService.isValidEmail(email));
-    }
-
-    @ParameterizedTest
-    @MethodSource("getLocaleArguments")
-    @DisplayName("getLocale() returns correct value")
-    void getLocale(String lang, Locale expected)
-    {
-        assertEquals(expected, localeService.getLocale(lang));
     }
 }
